Refactor with proper use of "pass"
I had a pass corresponding different frames, where it semms to be always
inside a frame from everything I have seen so far. I think I forgot/ got
confused with the terminology and concept (and was probably inspired to
not shift my code around, and then saw an early exit opportunity which led
to a better feel of swapping buffers, so turned out to be useful).