New standalone VS solution rendering simple triangle for geometry shader work
Goal of this commit was to get things compiling and rendering, so files were copied over from the linux/src folder and code resulting in compile time or runtime errors was stripped away. The solution was placed in the structure so that when copying the "Initial" solution, I wouldn't have to change lib and header paths.