Baseline tri and plane have textures rendered correctly again
Now that I only allocate the VBO once per model, I only had allocated enough memory for the pos - a previous optimisation I did. Textures were fine for the cube scene as I happened to have the same size allocated for both pos and tex (I did this because of the interleaved attributes of the cube). It is now better to simply assign the full size, as the optimisation of only sending over buffer data once per model to the GPU gives bigger gains.